Fix BinanceDataMining attributes. Files was restructured
