+++ custom_components/binance_pool/__init__.py
@@ -3,22 +3,15 @@ import logging
 import asyncio
 import copy
 
-from binance.client import AsyncClient
-from binance.exceptions import BinanceAPIException, BinanceRequestException
-
-import voluptuous as vol
-import homeassistant.helpers.config_validation as cv
-
 from homeassistant.const import CONF_API_KEY, CONF_NAME
 from homeassistant.helpers.discovery import async_load_platform
 from homeassistant.util import Throttle
 from homeassistant.helpers.update_coordinator import CoordinatorEntity, DataUpdateCoordinator, UpdateFailed
 from homeassistant.exceptions import ConfigEntryAuthFailed
+
+from .schemas import CONFIG_SCHEMA
 from .const import (
     DOMAIN,
-    DEFAULT_NAME,
-    DEFAULT_DOMAIN,
-    DEFAULT_CURRENCY,
     CONF_API_SECRET,
     CONF_BALANCES,
     CONF_EXCHANGES,
@@ -30,37 +23,13 @@ from .const import (
     COORDINATOR_MINING,
     COORDINATOR_WALLET    
 )
-__version__ = "1.5.6"
+
+from .client import BinancePoolClient, BinanceAPIException, BinanceRequestException
+
+__version__ = "1.5.7"
 REQUIREMENTS = ["python-binance==1.0.10"]
 
 _LOGGER = logging.getLogger(__name__)
-
-CONFIG_SCHEMA = vol.Schema(
-    {
-        DOMAIN: vol.Schema(
-            {
-                vol.Optional(CONF_NAME, default=DEFAULT_NAME): cv.string,
-                vol.Optional(CONF_DOMAIN, default=DEFAULT_DOMAIN): cv.string,
-                vol.Optional(CONF_NATIVE_CURRENCY, default=DEFAULT_CURRENCY): vol.All(
-                    cv.ensure_list, [cv.string]
-                ),
-                vol.Required(CONF_API_KEY): cv.string,
-                vol.Required(CONF_API_SECRET): cv.string,
-                vol.Optional(CONF_BALANCES, default=[]): vol.All(
-                    cv.ensure_list, [cv.string]
-                ),
-                vol.Optional(CONF_EXCHANGES, default=[]): vol.All(
-                    cv.ensure_list, [cv.string]
-                ),
-                vol.Optional(CONF_MINING, default=[]): vol.All(
-                    cv.ensure_list, [cv.string]
-                ),                
-            }
-        )
-    },
-    extra=vol.ALLOW_EXTRA,
-)
-
 
 async def async_setup(hass, config):
     api_key = config[DOMAIN][CONF_API_KEY]
@@ -74,6 +43,14 @@ async def async_setup(hass, config):
 
     binance_data_wallet = BinanceDataWallet(hass, api_key, api_secret, tld)
     binance_data_mining = BinanceDataMining(hass, api_key, api_secret, tld, miners)
+
+    hass.data[DOMAIN] = { 
+        'config': config,
+        'coordinator': {
+            COORDINATOR_MINING: binance_data_mining,
+            COORDINATOR_WALLET: binance_data_wallet
+        } 
+    }
     
     upddata = [ binance_data_wallet.async_config_entry_first_refresh() ]
     if miners:
@@ -89,13 +66,7 @@ async def async_setup(hass, config):
             await binance_data_mining.client.close_connection()
             raise r
 
-    hass.data[DOMAIN] = {
-        'config': config,
-        'coordinator': {
-             COORDINATOR_WALLET: binance_data_wallet,
-             COORDINATOR_MINING: binance_data_mining
-        }
-    }
+
     
     if hasattr(binance_data_wallet, "balances"):
         for balance in binance_data_wallet.balances:
@@ -212,7 +183,7 @@ async def async_setup(hass, config):
                     if "dayHashRate" not in status:
                         status["dayHashRate"] = 0                    
 
-                    for coindata in binance_data_wallet.coins:
+                    for coindata in binance_data_mining.coins:
                         if coindata["algoName"].lower() != algo:
                             continue
 
@@ -260,6 +231,7 @@ class BinanceDataMining(DataUpdateCoordinator):
         self.client = BinancePoolClient(api_key, api_secret, tld=tld)
         
         self.mining = {}
+        self.coins = {}
         self.tld = tld
         
         for account in miners:
@@ -315,7 +287,7 @@ class BinanceDataWallet(DataUpdateCoordinator):
         super().__init__(hass, _LOGGER, name="BinanceDataWallet", update_interval=timedelta(minutes=MIN_TIME_BETWEEN_UPDATES))
         
         self.client = BinancePoolClient(api_key, api_secret, tld=tld)
-        self.coins = {}
+        self.hass = hass
         self.balances = []
         self.funding = []
         self.savings = {}
@@ -358,7 +330,7 @@ class BinanceDataWallet(DataUpdateCoordinator):
             if prices:
                 self.tickers = prices
                 _LOGGER.debug(f"Exchange rates updated from binance.{self.tld}")
-            
+
             return True
         
         except (BinanceAPIException, BinanceRequestException) as e:
@@ -367,116 +339,3 @@ class BinanceDataWallet(DataUpdateCoordinator):
 
                                
             
-class BinancePoolClient(AsyncClient):
-    MINING_API_URL = 'https://api.binance.{}/sapi'
-    BALANCES_API_URL = 'https://api.binance.{}/sapi'
-    MINING_API_VERSION = 'v1'
-    BALANCES_API_VERSION = 'v1'
-    RECV_WINDOW = 50000
-    
-    def _get_request_kwargs(self, method, signed: bool, force_params: bool = False, **kwargs):
-        if signed:
-            kwargs['data']['recvWindow'] = self.RECV_WINDOW
-            
-        return super()._get_request_kwargs(method, signed, force_params, **kwargs)
-    
-    def _create_mining_api_url(self, path: str, version: str = MINING_API_URL ) -> str:
-        return self.MINING_API_URL.format(self.tld) + '/' + self.MINING_API_VERSION + '/mining/' + path        
-
-    def _create_capital_api_url(self, path: str, version: str = BALANCES_API_URL ) -> str:
-        return self.BALANCES_API_URL.format(self.tld) + '/' + self.BALANCES_API_VERSION + '/capital/' + path
-        
-    async def async_request_mining_api(self, method, path, signed=False, **kwargs):
-        uri = self._create_mining_api_url(path)
-        
-        answer = await self._request(method, uri, signed, True, **kwargs)
-        
-        if answer["code"] != 0 or "data" not in answer:
-            _LOGGER.error(f"Error fetching mining data from binance.{self.tld}: {answer}")
-            return False   
-             
-        return answer["data"]
-
-    async def async_request_capital_api(self, method, path, signed=False, **kwargs):
-        uri = self._create_capital_api_url(path)
-        
-        return await self._request(method, uri, signed, True, **kwargs)
-        
-    async def async_get_mining_algolist(self):
-        """ Acquiring Algorithm (MARKET_DATA)
-        
-            https://binance-docs.github.io/apidocs/spot/en/#acquiring-algorithm-market_data
-            
-        """
-        return await self.async_request_mining_api('get', 'pub/algoList')
-
-
-    async def async_get_mining_coinlist(self):
-        """ Acquiring CoinName (MARKET_DATA)
-        
-            https://binance-docs.github.io/apidocs/spot/en/#acquiring-coinname-market_data
-        """
-        return await self.async_request_mining_api('get', 'pub/coinList')        
-
-
-    async def async_get_mining_worker_detail(self, **params):
-        """ Request for Detail Miner List (USER_DATA)
-
-            https://binance-docs.github.io/apidocs/spot/en/#request-for-detail-miner-list-user_data
-        """
-        return await self.async_request_mining_api('get', 'worker/detail', True, data=params)        
-        
-
-    async def async_get_mining_worker_list(self, **params):
-        """ Request for Miner List (USER_DATA)
-
-            https://binance-docs.github.io/apidocs/spot/en/#earnings-list-user_data
-        """
-        return await self.async_request_mining_api('get', 'worker/list', True, data=params)        
-        
-    
-    async def async_get_mining_earning_history(self, **params):
-        """ Earnings List(USER_DATA)
-
-            https://binance-docs.github.io/apidocs/spot/en/#earnings-list-user_data
-        """
-        return await self.async_request_mining_api('get', 'payment/list', True, data=params)      
-
-
-    async def async_get_mining_bonus_history(self, **params):
-        """ Extra Bonus List (USER_DATA)
-
-            https://binance-docs.github.io/apidocs/spot/en/#extra-bonus-list-user_data
-        """
-        return await self.async_request_mining_api('get', 'payment/other', True, data=params) 
-
-
-    async def async_get_mining_status(self, **params):
-        """ Statistic List (USER_DATA)
-
-            https://binance-docs.github.io/apidocs/spot/en/#statistic-list-user_data
-        """
-        return await self.async_request_mining_api('get', 'statistics/user/status', True, data=params) 
-
-        
-    async def async_get_mining_history(self, **params):
-        """ Account List (USER_DATA)
-
-            https://binance-docs.github.io/apidocs/spot/en/#account-list-user_data
-        """
-        return await self.async_request_mining_api('get', 'statistics/user/list', True, data=params)           
-
-    async def async_get_capital_balances(self, **params):
-        """ All Coins' Information (USER_DATA)
-
-            https://binance-docs.github.io/apidocs/spot/en/#all-coins-39-information-user_data
-        """
-        return await self.async_request_capital_api('get', 'config/getall', True, data=params) 
-        
-    async def async_get_funding_balances(self, **params):
-        """ Funding Wallet (USER_DATA)
-        
-            https://binance-docs.github.io/apidocs/spot/en/#funding-wallet-user_data
-        """
-        return await self._request_margin_api('post', 'asset/get-funding-asset', True, data=params) 
-
