+++ custom_components/binance_pool/manifest.json
@@ -4,7 +4,7 @@
   "documentation": "https://github.com/shammysha/homeassistant-binance-pool",
   "issue_tracker": "https://github.com/shammysha/homeassistant-binance-pool/issues",
   "dependencies": [],
-  "version": "1.5.6",
+  "version": "1.5.7",
   "codeowners": [
     "@shammysha"
   ],
